Implements account verification check
Adds a component that checks if the selected account is verified
before allowing access to invoice features.

The component redirects users to the account verification page
if their account is not yet verified or if an error occurs
while retrieving the verification status.

Also refactors account not verified page to use clsx.